[FIX] make type and move to utils
